requirements.txt
pytest>=8.4.1
pytest-html>=4.1.1
deepeval>=3.3.3
python-dotenv>=1.1.1
requests>=2.32.4
matplotlib>=3.10.5
pandas>=2.3.1
pytest-xdist>=3.8.0
pytest-rerunfailures>=12.0
ollama>=0.5.1
playwright>=1.41.0
pytest-playwright>=0.4.4